refactor: split strategy chat prompt helpers
